fixed the pagePage view for when users view other user's pages

+++ FinalProject/app/views/Page/pagePage.php
@@ -14,10 +14,21 @@
                     } else {
                         echo "<a href=\"" . BASE . "/Star/delete/" . $data['page']->page_id . "\"><i class=\"fas fa-star\"></i></a>";
                     }
+
+
+                    if($_SESSION['profile_id'] !== $data['page']->profile_id){
+                        echo "<a href=\"" . BASE . "/Profile/viewProfile/" . $data['page']->profile_id . "\"> Go back </a>";
+                    }else{
+                        echo "<a href='" . BASE . "/Page/index/" . $_SESSSION['profile_id'] . "'> Return to page list </a>";
+                    }
+
                 ?>
-                <a href="<?= BASE ?>/Page/index/<?= $_SESSION['profile_id']?>"> Return to page list</a>
+       
                 <a href="<?= BASE ?>/Comment/add/<?= $data['page']->page_id?>"> Write comment</a> <br /> <br />
+
+
                 <a href="<?=BASE?>/Default/logout">Logout</a>
+
             </div>
 
             <div id="pageContents">
@@ -34,9 +45,17 @@
                 <br />
 
                 <?php 
-                    foreach ($data['comments'] as $comment) {
+                    if($_SESSION['profile_id'] !== $data['page']->profile_id){
+                        foreach ($data['comments'] as $comment) {
+                        echo "$comment->comment_text <br />";
+                        }
+                    }else{
+                        foreach ($data['comments'] as $comment) {
                         echo "$comment->comment_text <a href=\"" . BASE . "/Comment/delete/" . $comment->comment_id . "\">delete</a> <br />";
+                        }
                     }
+
+                    
                 ?>
             </div>
         </div>
